refactor(GameBoard) : 점수 계산 participant로 이전
