+++ src/main/java/domain/participant/Player.java
@@ -1,17 +1,8 @@
 package domain.participant;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class Player implements Participant {
-    private static final int BUST_THRESHOLD = 21;
-
-    private final String nickname;
-    private final Map<BattleResult, Integer> battleResult;
-
+public class Player extends Participant {
     private Player(final String nickname) {
-        this.nickname = nickname;
-        this.battleResult = new HashMap<>();
+        super(nickname);
     }
 
     public static Player from(final String nickname) {
@@ -26,20 +17,5 @@ public class Player implements Participant {
     @Override
     public boolean areYouDealer() {
         return false;
-    }
-
-    @Override
-    public String getNickname() {
-        return nickname;
-    }
-
-    @Override
-    public Map<BattleResult, Integer> getBattleResult() {
-        return battleResult;
-    }
-
-    @Override
-    public void updateBattleResult(final BattleResult battleResult) {
-        this.battleResult.merge(battleResult, 1, Integer::sum);
     }
 }
